+++ src/domain/common/enums/Category.ts
